Update - Move manual sampler control methods to samplerSelect.js
