+++ public/scripts/textgen-settings.js
@@ -20,13 +20,13 @@ import { autoSelectInstructPreset, selectContextPreset, selectInstructPreset } f
 import { BIAS_CACHE, createNewLogitBiasEntry, displayLogitBias, getLogitBiasListResult } from './logit-bias.js';
 
 import { power_user, registerDebugFunction } from './power-user.js';
+import { getManualActivePresetSamplers, isSamplerManualPriorityEnabled, loadPresetSelectedSamplers } from './samplerSelect.js';
 import { SECRET_KEYS, writeSecret } from './secrets.js';
 import { getEventSourceStream } from './sse-stream.js';
 import { getCurrentDreamGenModelTokenizer, getCurrentOpenRouterModelTokenizer, loadAphroditeModels, loadDreamGenModels, loadFeatherlessModels, loadGenericModels, loadInfermaticAIModels, loadMancerModels, loadOllamaModels, loadOpenRouterModels, loadTabbyModels, loadTogetherAIModels, loadVllmModels } from './textgen-models.js';
 import { ENCODE_TOKENIZERS, TEXTGEN_TOKENIZERS, TOKENIZER_SUPPORTED_KEY, getTextTokens, tokenizers } from './tokenizers.js';
 import { AbortReason } from './util/AbortReason.js';
 import { getSortableDelay, onlyUnique, arraysEqual, isObject } from './utils.js';
-import { localforage } from '../lib.js';
 
 export const textgen_types = {
     OOBA: 'ooba',
@@ -114,9 +114,6 @@ export const APHRODITE_DEFAULT_ORDER = [
     'xtc',
 ];
 const BIAS_KEY = '#textgenerationwebui_api-settings';
-
-const textGenObjectStore = localforage.createInstance({ name: 'SillyTavern_TextCompletions' });
-let selectedSamplers = {};
 
 // Maybe let it be configurable in the future?
 // (7 days later) The future has come.
@@ -360,123 +357,6 @@ export function getTextGenServer(type = null) {
         default:
             return settings.server_urls[selectedType] ?? '';
     }
-}
-
-/**
- * Initializes the configuration object for manually selected samplers.
- */
-async function loadPresetSelectedSamplers() {
-    try {
-        console.debug('Text Completions: loading selected samplers');
-        selectedSamplers = await textGenObjectStore.getItem('selectedSamplers') || {};
-    } catch (error) {
-        console.log('Text Completions: unable to load selected samplers, using default samplers', error);
-        selectedSamplers = {};
-    }
-}
-
-/**
- * Synchronizes the local forage instance with the selected samplers configuration object.
- */
-export async function savePresetSelectedSamplers() {
-    try {
-        console.debug('Text Completions: saving selected samplers');
-        await textGenObjectStore.setItem('selectedSamplers', selectedSamplers);
-    } catch (error) {
-        console.log('Text Completions: unable to save selected samplers', error);
-    }
-}
-
-/**
- * Resets the selected samplers configuration object from the local forage instace.
- * @param {string?} preset_name Name of the target preset - It picks the current active TC preset name by default
- * @param {boolean} silent Suppresses the toastr message confirming that the data was deleted.
- */
-export async function resetPresetSelectedSamplers(preset_name = '', silent = false) {
-    try {
-        if (!settings?.preset && !preset_name) return;
-        if (!preset_name) preset_name = settings.preset;
-        if (!selectedSamplers[preset_name]) return;
-
-        console.debug('Text Completions: resetting selected samplers');
-        delete selectedSamplers[preset_name];
-        await savePresetSelectedSamplers();
-        if (!silent) toastr.success('Selected samplers cleared.');
-    } catch (error) {
-        console.log('Text Completions: unable to reset selected preset samplers', error);
-    }
-}
-
-/**
- * Saves the visibility state for selected samplers into the configuration object.
- * @param {string} sampler_name Target sampler key name
- * @param {string|boolean} state Visibility state of the target sampler
- * @param {string?} preset_name Name of the target preset - It picks the current active TC preset name by default
- * @returns void
- */
-export function setPresetSamplersState(sampler_name, state, preset_name = '') {
-    if (!settings?.preset && !preset_name) return;
-    if (!preset_name) preset_name = settings.preset;
-    if (!selectedSamplers[preset_name]) selectedSamplers[preset_name] = {};
-
-    const presetSamplers = selectedSamplers[preset_name];
-    presetSamplers[sampler_name] = String(state) === 'true';
-}
-
-export function getManualPresetSamplers(preset_name = '') {
-    if (!settings?.preset && !preset_name) return {};
-    if (!preset_name) preset_name = settings.preset;
-    if (!selectedSamplers[preset_name]) selectedSamplers[preset_name] = {};
-
-    return selectedSamplers[preset_name];
-}
-
-/**
- * Returns the key names of all the preset samplers activated manually.
- * @param {string?} preset_name Name of the target preset - It picks the current active TC preset name by default
- * @returns {string[]} Array of sampler key names
- */
-export function getManualActivePresetSamplers(preset_name = '') {
-    if (!settings?.preset && !preset_name) return [];
-    if (!preset_name) preset_name = settings.preset;
-    if (!selectedSamplers[preset_name]) selectedSamplers[preset_name] = {};
-
-    try {
-        const presetSamplers = Object.entries(selectedSamplers[preset_name]);
-
-        return presetSamplers
-            .filter(([key, val]) => val === true && key !== 'st_manual_priority')
-            .map(([key, val]) => key);
-    } catch (error) {
-        console.log('Text Completions: unable to fetch active preset samplers', error);
-        return [];
-    }
-}
-
-/**
- * @param {string|boolean} state Target state of the feature
- * @param {string?} preset_name Name of the target preset - It picks the current active TC preset name by default
- * @returns void
- */
-export function toggleSamplerManualPriority(state = false, preset_name = '') {
-    if (!settings?.preset && !preset_name) return;
-    if (!preset_name) preset_name = settings.preset;
-    if (!selectedSamplers[preset_name]) selectedSamplers[preset_name] = {};
-
-    const presetSamplers = selectedSamplers[preset_name];
-    presetSamplers.st_manual_priority = String(state) === 'true';
-}
-
-/**
- * @param {string?} preset_name Name of the target preset - It picks the current active TC preset name by default
- * @returns {boolean}
- */
-export function isSamplerManualPriorityEnabled(preset_name = '') {
-    if (!settings?.preset && !preset_name) return false;
-    if (!preset_name) preset_name = settings.preset;
-    if (!selectedSamplers[preset_name]) selectedSamplers[preset_name] = {};
-
-    return selectedSamplers[preset_name]?.st_manual_priority ?? false;
 }
 
 async function selectPreset(name) {
